Rename `inline_simple_functions` to `inline_functions_with_at_most_one_instruction`
